+++ api/src/main/java/org/allaymc/api/eventbus/event/entity/ProjectileLaunchEvent.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.eventbus.event.entity;
 
 import lombok.Getter;
+import lombok.Setter;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.interfaces.EntityProjectile;
 import org.allaymc.api.eventbus.event.CancellableEvent;
@@ -15,10 +16,13 @@ public class ProjectileLaunchEvent extends EntityEvent implements CancellableEve
      * The entity that shoot this projectile.
      */
     protected Entity shooter;
+    @Setter
+    protected double throwForce;
 
-    public ProjectileLaunchEvent(EntityProjectile entity, Entity shooter) {
+    public ProjectileLaunchEvent(EntityProjectile entity, Entity shooter, double throwForce) {
         super(entity);
         this.shooter = shooter;
+        this.throwForce = throwForce;
     }
 
     @Override